test(web): extract shared visualizer synthetic-source story helper
