+++ apps/web/src/components/player/VuVisualizer/VuVisualizer.stories.tsx
@@ -2,31 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import { expect } from 'storybook/test';
 
 import VuVisualizer from './VuVisualizer';
-import type { FrequencySource } from '../visualizer-source';
-
-/**
- * A deterministic, audio-free frequency source so a story can drive the real
- * per-frame draw path without the playback engine — two summed sine waves
- * tapered toward the high bins, mirroring the rough shape of real spectrum
- * data (same shape the settings preview uses).
- */
-function createSyntheticSource(): FrequencySource {
-  let t = 0;
-  return {
-    binCount: 256,
-    read(buf) {
-      t += 0.04;
-      for (let i = 0; i < buf.length; i++) {
-        const x = i / buf.length;
-        const a = Math.sin(t + x * 7) * (1 - x);
-        const b = Math.sin(t * 1.6 + x * 3.5 + 1.1) * (1 - x * 0.55);
-        buf[i] = Math.max(0, Math.min(255, ((a + b) * 0.5 + 0.5) * 255 * 0.65));
-      }
-      return true;
-    },
-  };
-}
-
+import { createSyntheticSource } from '../visualizerStorySource';
 /**
  * player · VuVisualizer. Twin analog VU needles (L/R) sweeping over a ticked
  * arc. Renders a single decorative `<canvas>` (no role, no text) that animates
